+++ banyand/cmd/dump/trace.go
@@ -32,8 +32,8 @@ import (
 	"github.com/apache/skywalking-banyandb/api/common"
 	databasev1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/database/v1"
 	modelv1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/model/v1"
-	"github.com/apache/skywalking-banyandb/banyand/dump"
-	dumptrace "github.com/apache/skywalking-banyandb/banyand/dump/trace"
+	"github.com/apache/skywalking-banyandb/banyand/internal/dump"
+	dumptrace "github.com/apache/skywalking-banyandb/banyand/internal/dump/trace"
 	"github.com/apache/skywalking-banyandb/pkg/convert"
 	"github.com/apache/skywalking-banyandb/pkg/encoding"
 	"github.com/apache/skywalking-banyandb/pkg/fs"
