+++ banyand/internal/dump/measure/suite_test.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package trace
+package measure
 
 import (
 	"context"
@@ -33,14 +33,14 @@ import (
 	"github.com/apache/skywalking-banyandb/api/data"
 	commonv1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/common/v1"
 	databasev1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/database/v1"
+	measurev1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/measure/v1"
 	modelv1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/model/v1"
-	tracev1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/trace/v1"
-	"github.com/apache/skywalking-banyandb/banyand/dump"
+	"github.com/apache/skywalking-banyandb/banyand/internal/dump"
+	"github.com/apache/skywalking-banyandb/banyand/measure"
 	metadataservice "github.com/apache/skywalking-banyandb/banyand/metadata/service"
 	obsservice "github.com/apache/skywalking-banyandb/banyand/observability/services"
 	"github.com/apache/skywalking-banyandb/banyand/protector"
 	"github.com/apache/skywalking-banyandb/banyand/queue"
-	"github.com/apache/skywalking-banyandb/banyand/trace"
 	"github.com/apache/skywalking-banyandb/pkg/bus"
 	localfs "github.com/apache/skywalking-banyandb/pkg/fs"
 	"github.com/apache/skywalking-banyandb/pkg/logger"
@@ -48,16 +48,17 @@ import (
 )
 
 const (
-	rtGroup = "dump_rt_group"
-	rtTrace = "dump_rt_trace"
+	rtGroup   = "dump_rt_group"
+	rtMeasure = "dump_rt_measure"
 )
 
-// TestTraceWriteRoundTrip drives the top-level trace write path (the same
+// TestMeasureWriteRoundTrip drives the top-level measure write path (the same
 // writeCallback.Rev reached via the data pipeline), flushes real parts to disk,
 // then reads them back with the dump iterator and verifies every written field.
-func TestTraceWriteRoundTrip(t *testing.T) {
+func TestMeasureWriteRoundTrip(t *testing.T) {
 	req := require.New(t)
 	require.NoError(t, logger.Init(logger.Logging{Env: "dev", Level: "warn"}))
+	// Bridge gomega (used inside test.SetupModules) onto the test process.
 	gomega.RegisterFailHandler(func(message string, _ ...int) { panic(message) })
 
 	pipeline := queue.Local()
@@ -65,7 +66,7 @@ func TestTraceWriteRoundTrip(t *testing.T) {
 	req.NoError(err)
 	metricSvc := obsservice.NewMetricService(metaSvc, pipeline, "test", nil)
 	pm := protector.NewMemory(metricSvc)
-	traceSvc, err := trace.NewService(metaSvc, pipeline, metricSvc, pm)
+	measureSvc, err := measure.NewStandalone(metaSvc, pipeline, nil, metricSvc, pm)
 	req.NoError(err)
 
 	metaPath, metaDefer, err := test.NewSpace()
@@ -81,29 +82,29 @@ func TestTraceWriteRoundTrip(t *testing.T) {
 		"--schema-server-root-path=" + metaPath,
 		fmt.Sprintf("--schema-server-grpc-port=%d", ports[0]),
 		"--schema-server-grpc-host=127.0.0.1",
-		"--trace-root-path=" + rootPath,
-		"--trace-flush-timeout=200ms",
+		"--measure-root-path=" + rootPath,
+		"--measure-flush-timeout=200ms",
 	}
-	moduleDefer := test.SetupModules(flags, pipeline, metaSvc, traceSvc)
+	moduleDefer := test.SetupModules(flags, pipeline, metaSvc, measureSvc)
 	defer moduleDefer()
 
-	registerRoundTripTrace(t, metaSvc)
+	registerRoundTripMeasure(t, metaSvc)
 	require.Eventually(t, func() bool {
-		_, ok := traceSvc.LoadGroup(rtGroup)
+		_, ok := measureSvc.LoadGroup(rtGroup)
 		return ok
-	}, 30*time.Second, 200*time.Millisecond, "trace group not loaded")
+	}, 30*time.Second, 200*time.Millisecond, "measure group not loaded")
+	// Let the measure schema event propagate after the group is ready.
 	time.Sleep(time.Second)
 
-	tagNames := []string{"trace_id", "span_id", "timestamp", "service_id", "duration"}
 	const total = 3
 	baseTS := time.Now().Truncate(time.Minute)
 	type want struct {
-		traceID  string
-		spanID   string
-		service  string
-		span     string
-		duration int64
-		tsMilli  int64
+		series  string
+		strTag  string
+		intTag  int64
+		intF    int64
+		floatF  float64
+		tsMilli int64
 	}
 	wants := map[string]want{}
 
@@ -112,29 +113,34 @@ func TestTraceWriteRoundTrip(t *testing.T) {
 		iStr := strconv.Itoa(i)
 		ts := baseTS.Add(time.Duration(i) * time.Minute)
 		w := want{
-			traceID:  "trace" + iStr,
-			spanID:   "span" + iStr,
-			service:  "svc" + iStr,
-			duration: int64(100 + i),
-			span:     "span_data_" + iStr,
-			tsMilli:  ts.UnixMilli(),
+			series:  "series" + iStr,
+			strTag:  "str" + iStr,
+			intTag:  int64(10 + i),
+			intF:    int64(100 + i),
+			floatF:  float64(i) + 0.5,
+			tsMilli: ts.UnixMilli(),
 		}
-		wants[w.spanID] = w
-		tags := []*modelv1.TagValue{
-			{Value: &modelv1.TagValue_Str{Str: &modelv1.Str{Value: w.traceID}}},
-			{Value: &modelv1.TagValue_Str{Str: &modelv1.Str{Value: w.spanID}}},
-			{Value: &modelv1.TagValue_Timestamp{Timestamp: timestamppb.New(ts)}},
-			{Value: &modelv1.TagValue_Str{Str: &modelv1.Str{Value: w.service}}},
-			{Value: &modelv1.TagValue_Int{Int: &modelv1.Int{Value: w.duration}}},
+		wants[w.strTag] = w
+		writeReq := &measurev1.WriteRequest{
+			Metadata: &commonv1.Metadata{Name: rtMeasure, Group: rtGroup},
+			DataPoint: &measurev1.DataPointValue{
+				Timestamp: timestamppb.New(ts),
+				TagFamilies: []*modelv1.TagFamilyForWrite{{
+					Tags: []*modelv1.TagValue{
+						{Value: &modelv1.TagValue_Str{Str: &modelv1.Str{Value: w.series}}},
+						{Value: &modelv1.TagValue_Str{Str: &modelv1.Str{Value: w.strTag}}},
+						{Value: &modelv1.TagValue_Int{Int: &modelv1.Int{Value: w.intTag}}},
+					},
+				}},
+				Fields: []*modelv1.FieldValue{
+					{Value: &modelv1.FieldValue_Int{Int: &modelv1.Int{Value: w.intF}}},
+					{Value: &modelv1.FieldValue_Float{Float: &modelv1.Float{Value: w.floatF}}},
+				},
+			},
 		}
-		writeReq := &tracev1.WriteRequest{
-			Metadata: &commonv1.Metadata{Name: rtTrace, Group: rtGroup},
-			Tags:     tags,
-			Span:     []byte(w.span),
-			TagSpec:  &tracev1.TagSpec{TagNames: tagNames},
-		}
-		_, errPub := bp.Publish(context.TODO(), data.TopicTraceWrite, bus.NewMessage(bus.MessageID(i), &tracev1.InternalWriteRequest{
-			Request: writeReq,
+		_, errPub := bp.Publish(context.TODO(), data.TopicMeasureWrite, bus.NewMessage(bus.MessageID(i), &measurev1.InternalWriteRequest{
+			EntityValues: []*modelv1.TagValue{{Value: &modelv1.TagValue_Str{Str: &modelv1.Str{Value: w.series}}}},
+			Request:      writeReq,
 		}))
 		req.NoError(errPub)
 	}
@@ -157,33 +163,35 @@ func TestTraceWriteRoundTrip(t *testing.T) {
 		it := p.Iterator()
 		for it.Next() {
 			r := it.Row()
-			w, ok := wants[r.SpanID]
-			require.True(t, ok, "unexpected row spanID=%q", r.SpanID)
-			require.False(t, seen[r.SpanID], "duplicate row for %s", r.SpanID)
-			seen[r.SpanID] = true
+			strTag := dump.DecodeTagValue(r.TagTypes["default.strTag"], r.Tags["default.strTag"], nil).GetStr().GetValue()
+			w, ok := wants[strTag]
+			require.True(t, ok, "unexpected row strTag=%q sid=%d", strTag, r.SeriesID)
+			require.False(t, seen[w.strTag], "duplicate row for %s", w.strTag)
+			seen[w.strTag] = true
 			seriesIDs[uint64(r.SeriesID)] = true
 
-			require.Equal(t, w.traceID, r.TraceID)
-			require.Equal(t, w.span, string(r.Span))
-			require.Equal(t, w.service, dump.DecodeTagValue(r.TagTypes["service_id"], r.Tags["service_id"], nil).GetStr().GetValue())
-			require.Equal(t, w.duration, dump.DecodeTagValue(r.TagTypes["duration"], r.Tags["duration"], nil).GetInt().GetValue())
-			require.Equal(t, w.tsMilli, time.Unix(0, r.Timestamp).UnixMilli(), "timestamp round-trip for %s", r.SpanID)
-			require.NotZero(t, r.SeriesID, "seriesID derived from tags")
+			require.Equal(t, w.intTag, dump.DecodeTagValue(r.TagTypes["default.intTag"], r.Tags["default.intTag"], nil).GetInt().GetValue())
+			require.Equal(t, w.intF, DecodeFieldValue(r.FieldTypes["intField"], r.Fields["intField"]).GetInt().GetValue())
+			require.InDelta(t, w.floatF, DecodeFieldValue(r.FieldTypes["floatField"], r.Fields["floatField"]).GetFloat().GetValue(), 1e-9)
+			require.Equal(t, w.tsMilli, time.Unix(0, r.Timestamp).UnixMilli(), "timestamp round-trip for %s", w.strTag)
+			require.NotZero(t, r.SeriesID, "seriesID derived from entity")
+			// The entity tag itself is not stored as a data column.
+			require.NotContains(t, r.Tags, "default.series")
 		}
 		require.NoError(t, it.Err())
 		it.Close()
 		p.Close()
 	}
-	require.Len(t, seen, total, "every written span must be read back exactly once")
-	require.Len(t, seriesIDs, total, "each span must map to a distinct series")
+	require.Len(t, seen, total, "every written data point must be read back exactly once")
+	require.Len(t, seriesIDs, total, "each distinct entity must map to a distinct series")
 }
 
-func registerRoundTripTrace(t *testing.T, metaSvc metadataservice.Service) {
+func registerRoundTripMeasure(t *testing.T, metaSvc metadataservice.Service) {
 	reg := metaSvc.SchemaRegistry()
 	ctx := context.TODO()
 	_, err := reg.CreateGroup(ctx, &commonv1.Group{
 		Metadata: &commonv1.Metadata{Name: rtGroup},
-		Catalog:  commonv1.Catalog_CATALOG_TRACE,
+		Catalog:  commonv1.Catalog_CATALOG_MEASURE,
 		ResourceOpts: &commonv1.ResourceOpts{
 			ShardNum:        1,
 			SegmentInterval: &commonv1.IntervalRule{Unit: commonv1.IntervalRule_UNIT_DAY, Num: 1},
@@ -191,22 +199,37 @@ func registerRoundTripTrace(t *testing.T, metaSvc metadataservice.Service) {
 		},
 	})
 	require.NoError(t, err)
-	_, err = reg.CreateTrace(ctx, &databasev1.Trace{
-		Metadata: &commonv1.Metadata{Name: rtTrace, Group: rtGroup},
-		Tags: []*databasev1.TraceTagSpec{
-			{Name: "trace_id", Type: databasev1.TagType_TAG_TYPE_STRING},
-			{Name: "span_id", Type: databasev1.TagType_TAG_TYPE_STRING},
-			{Name: "timestamp", Type: databasev1.TagType_TAG_TYPE_TIMESTAMP},
-			{Name: "service_id", Type: databasev1.TagType_TAG_TYPE_STRING},
-			{Name: "duration", Type: databasev1.TagType_TAG_TYPE_INT},
+	_, err = reg.CreateMeasure(ctx, &databasev1.Measure{
+		Metadata: &commonv1.Metadata{Name: rtMeasure, Group: rtGroup},
+		TagFamilies: []*databasev1.TagFamilySpec{{
+			Name: "default",
+			Tags: []*databasev1.TagSpec{
+				{Name: "series", Type: databasev1.TagType_TAG_TYPE_STRING},
+				{Name: "strTag", Type: databasev1.TagType_TAG_TYPE_STRING},
+				{Name: "intTag", Type: databasev1.TagType_TAG_TYPE_INT},
+			},
+		}},
+		Fields: []*databasev1.FieldSpec{
+			{
+				Name:              "intField",
+				FieldType:         databasev1.FieldType_FIELD_TYPE_INT,
+				EncodingMethod:    databasev1.EncodingMethod_ENCODING_METHOD_GORILLA,
+				CompressionMethod: databasev1.CompressionMethod_COMPRESSION_METHOD_ZSTD,
+			},
+			{
+				Name:              "floatField",
+				FieldType:         databasev1.FieldType_FIELD_TYPE_FLOAT,
+				EncodingMethod:    databasev1.EncodingMethod_ENCODING_METHOD_GORILLA,
+				CompressionMethod: databasev1.CompressionMethod_COMPRESSION_METHOD_ZSTD,
+			},
 		},
-		TraceIdTagName:   "trace_id",
-		SpanIdTagName:    "span_id",
-		TimestampTagName: "timestamp",
+		Entity: &databasev1.Entity{TagNames: []string{"series"}},
 	})
 	require.NoError(t, err)
 }
 
+// findPartDirs walks root and returns every directory that directly contains a
+// metadata.json file and whose name is a 16-hex part id (an on-disk part dir).
 func findPartDirs(root string) []string {
 	var dirs []string
 	_ = filepath.WalkDir(root, func(path string, d fs.DirEntry, err error) error {
