+++ banyand/internal/dump/stream/suite_test.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package measure
+package stream
 
 import (
 	"context"
@@ -33,14 +33,14 @@ import (
 	"github.com/apache/skywalking-banyandb/api/data"
 	commonv1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/common/v1"
 	databasev1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/database/v1"
-	measurev1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/measure/v1"
 	modelv1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/model/v1"
-	"github.com/apache/skywalking-banyandb/banyand/dump"
-	"github.com/apache/skywalking-banyandb/banyand/measure"
+	streamv1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/stream/v1"
+	"github.com/apache/skywalking-banyandb/banyand/internal/dump"
 	metadataservice "github.com/apache/skywalking-banyandb/banyand/metadata/service"
 	obsservice "github.com/apache/skywalking-banyandb/banyand/observability/services"
 	"github.com/apache/skywalking-banyandb/banyand/protector"
 	"github.com/apache/skywalking-banyandb/banyand/queue"
+	"github.com/apache/skywalking-banyandb/banyand/stream"
 	"github.com/apache/skywalking-banyandb/pkg/bus"
 	localfs "github.com/apache/skywalking-banyandb/pkg/fs"
 	"github.com/apache/skywalking-banyandb/pkg/logger"
@@ -48,17 +48,16 @@ import (
 )
 
 const (
-	rtGroup   = "dump_rt_group"
-	rtMeasure = "dump_rt_measure"
+	rtGroup  = "dump_rt_group"
+	rtStream = "dump_rt_stream"
 )
 
-// TestMeasureWriteRoundTrip drives the top-level measure write path (the same
+// TestStreamWriteRoundTrip drives the top-level stream write path (the same
 // writeCallback.Rev reached via the data pipeline), flushes real parts to disk,
 // then reads them back with the dump iterator and verifies every written field.
-func TestMeasureWriteRoundTrip(t *testing.T) {
+func TestStreamWriteRoundTrip(t *testing.T) {
 	req := require.New(t)
 	require.NoError(t, logger.Init(logger.Logging{Env: "dev", Level: "warn"}))
-	// Bridge gomega (used inside test.SetupModules) onto the test process.
 	gomega.RegisterFailHandler(func(message string, _ ...int) { panic(message) })
 
 	pipeline := queue.Local()
@@ -66,7 +65,7 @@ func TestMeasureWriteRoundTrip(t *testing.T) {
 	req.NoError(err)
 	metricSvc := obsservice.NewMetricService(metaSvc, pipeline, "test", nil)
 	pm := protector.NewMemory(metricSvc)
-	measureSvc, err := measure.NewStandalone(metaSvc, pipeline, nil, metricSvc, pm)
+	streamSvc, err := stream.NewService(metaSvc, pipeline, metricSvc, pm, nil)
 	req.NoError(err)
 
 	metaPath, metaDefer, err := test.NewSpace()
@@ -82,18 +81,17 @@ func TestMeasureWriteRoundTrip(t *testing.T) {
 		"--schema-server-root-path=" + metaPath,
 		fmt.Sprintf("--schema-server-grpc-port=%d", ports[0]),
 		"--schema-server-grpc-host=127.0.0.1",
-		"--measure-root-path=" + rootPath,
-		"--measure-flush-timeout=200ms",
+		"--stream-root-path=" + rootPath,
+		"--stream-flush-timeout=200ms",
 	}
-	moduleDefer := test.SetupModules(flags, pipeline, metaSvc, measureSvc)
+	moduleDefer := test.SetupModules(flags, pipeline, metaSvc, streamSvc)
 	defer moduleDefer()
 
-	registerRoundTripMeasure(t, metaSvc)
+	registerRoundTripStream(t, metaSvc)
 	require.Eventually(t, func() bool {
-		_, ok := measureSvc.LoadGroup(rtGroup)
+		_, ok := streamSvc.LoadGroup(rtGroup)
 		return ok
-	}, 30*time.Second, 200*time.Millisecond, "measure group not loaded")
-	// Let the measure schema event propagate after the group is ready.
+	}, 30*time.Second, 200*time.Millisecond, "stream group not loaded")
 	time.Sleep(time.Second)
 
 	const total = 3
@@ -102,8 +100,6 @@ func TestMeasureWriteRoundTrip(t *testing.T) {
 		series  string
 		strTag  string
 		intTag  int64
-		intF    int64
-		floatF  float64
 		tsMilli int64
 	}
 	wants := map[string]want{}
@@ -112,19 +108,13 @@ func TestMeasureWriteRoundTrip(t *testing.T) {
 	for i := 0; i < total; i++ {
 		iStr := strconv.Itoa(i)
 		ts := baseTS.Add(time.Duration(i) * time.Minute)
-		w := want{
-			series:  "series" + iStr,
-			strTag:  "str" + iStr,
-			intTag:  int64(10 + i),
-			intF:    int64(100 + i),
-			floatF:  float64(i) + 0.5,
-			tsMilli: ts.UnixMilli(),
-		}
+		w := want{series: "series" + iStr, strTag: "str" + iStr, intTag: int64(10 + i), tsMilli: ts.UnixMilli()}
 		wants[w.strTag] = w
-		writeReq := &measurev1.WriteRequest{
-			Metadata: &commonv1.Metadata{Name: rtMeasure, Group: rtGroup},
-			DataPoint: &measurev1.DataPointValue{
+		writeReq := &streamv1.WriteRequest{
+			Metadata: &commonv1.Metadata{Name: rtStream, Group: rtGroup},
+			Element: &streamv1.ElementValue{
 				Timestamp: timestamppb.New(ts),
+				ElementId: "element" + iStr,
 				TagFamilies: []*modelv1.TagFamilyForWrite{{
 					Tags: []*modelv1.TagValue{
 						{Value: &modelv1.TagValue_Str{Str: &modelv1.Str{Value: w.series}}},
@@ -132,13 +122,9 @@ func TestMeasureWriteRoundTrip(t *testing.T) {
 						{Value: &modelv1.TagValue_Int{Int: &modelv1.Int{Value: w.intTag}}},
 					},
 				}},
-				Fields: []*modelv1.FieldValue{
-					{Value: &modelv1.FieldValue_Int{Int: &modelv1.Int{Value: w.intF}}},
-					{Value: &modelv1.FieldValue_Float{Float: &modelv1.Float{Value: w.floatF}}},
-				},
 			},
 		}
-		_, errPub := bp.Publish(context.TODO(), data.TopicMeasureWrite, bus.NewMessage(bus.MessageID(i), &measurev1.InternalWriteRequest{
+		_, errPub := bp.Publish(context.TODO(), data.TopicStreamWrite, bus.NewMessage(bus.MessageID(i), &streamv1.InternalWriteRequest{
 			EntityValues: []*modelv1.TagValue{{Value: &modelv1.TagValue_Str{Str: &modelv1.Str{Value: w.series}}}},
 			Request:      writeReq,
 		}))
@@ -155,6 +141,7 @@ func TestMeasureWriteRoundTrip(t *testing.T) {
 	fileSystem := localfs.NewLocalFileSystem()
 	seen := map[string]bool{}
 	seriesIDs := map[uint64]bool{}
+	elementIDs := map[uint64]bool{}
 	for _, partDir := range partDirs {
 		partID, parseErr := strconv.ParseUint(filepath.Base(partDir), 16, 64)
 		req.NoError(parseErr)
@@ -169,29 +156,29 @@ func TestMeasureWriteRoundTrip(t *testing.T) {
 			require.False(t, seen[w.strTag], "duplicate row for %s", w.strTag)
 			seen[w.strTag] = true
 			seriesIDs[uint64(r.SeriesID)] = true
+			elementIDs[r.ElementID] = true
 
 			require.Equal(t, w.intTag, dump.DecodeTagValue(r.TagTypes["default.intTag"], r.Tags["default.intTag"], nil).GetInt().GetValue())
-			require.Equal(t, w.intF, DecodeFieldValue(r.FieldTypes["intField"], r.Fields["intField"]).GetInt().GetValue())
-			require.InDelta(t, w.floatF, DecodeFieldValue(r.FieldTypes["floatField"], r.Fields["floatField"]).GetFloat().GetValue(), 1e-9)
 			require.Equal(t, w.tsMilli, time.Unix(0, r.Timestamp).UnixMilli(), "timestamp round-trip for %s", w.strTag)
 			require.NotZero(t, r.SeriesID, "seriesID derived from entity")
-			// The entity tag itself is not stored as a data column.
+			require.NotZero(t, r.ElementID, "elementID set")
 			require.NotContains(t, r.Tags, "default.series")
 		}
 		require.NoError(t, it.Err())
 		it.Close()
 		p.Close()
 	}
-	require.Len(t, seen, total, "every written data point must be read back exactly once")
+	require.Len(t, seen, total, "every written element must be read back exactly once")
 	require.Len(t, seriesIDs, total, "each distinct entity must map to a distinct series")
+	require.Len(t, elementIDs, total, "each element must have a distinct elementID")
 }
 
-func registerRoundTripMeasure(t *testing.T, metaSvc metadataservice.Service) {
+func registerRoundTripStream(t *testing.T, metaSvc metadataservice.Service) {
 	reg := metaSvc.SchemaRegistry()
 	ctx := context.TODO()
 	_, err := reg.CreateGroup(ctx, &commonv1.Group{
 		Metadata: &commonv1.Metadata{Name: rtGroup},
-		Catalog:  commonv1.Catalog_CATALOG_MEASURE,
+		Catalog:  commonv1.Catalog_CATALOG_STREAM,
 		ResourceOpts: &commonv1.ResourceOpts{
 			ShardNum:        1,
 			SegmentInterval: &commonv1.IntervalRule{Unit: commonv1.IntervalRule_UNIT_DAY, Num: 1},
@@ -199,8 +186,8 @@ func registerRoundTripMeasure(t *testing.T, metaSvc metadataservice.Service) {
 		},
 	})
 	require.NoError(t, err)
-	_, err = reg.CreateMeasure(ctx, &databasev1.Measure{
-		Metadata: &commonv1.Metadata{Name: rtMeasure, Group: rtGroup},
+	_, err = reg.CreateStream(ctx, &databasev1.Stream{
+		Metadata: &commonv1.Metadata{Name: rtStream, Group: rtGroup},
 		TagFamilies: []*databasev1.TagFamilySpec{{
 			Name: "default",
 			Tags: []*databasev1.TagSpec{
@@ -209,27 +196,11 @@ func registerRoundTripMeasure(t *testing.T, metaSvc metadataservice.Service) {
 				{Name: "intTag", Type: databasev1.TagType_TAG_TYPE_INT},
 			},
 		}},
-		Fields: []*databasev1.FieldSpec{
-			{
-				Name:              "intField",
-				FieldType:         databasev1.FieldType_FIELD_TYPE_INT,
-				EncodingMethod:    databasev1.EncodingMethod_ENCODING_METHOD_GORILLA,
-				CompressionMethod: databasev1.CompressionMethod_COMPRESSION_METHOD_ZSTD,
-			},
-			{
-				Name:              "floatField",
-				FieldType:         databasev1.FieldType_FIELD_TYPE_FLOAT,
-				EncodingMethod:    databasev1.EncodingMethod_ENCODING_METHOD_GORILLA,
-				CompressionMethod: databasev1.CompressionMethod_COMPRESSION_METHOD_ZSTD,
-			},
-		},
 		Entity: &databasev1.Entity{TagNames: []string{"series"}},
 	})
 	require.NoError(t, err)
 }
 
-// findPartDirs walks root and returns every directory that directly contains a
-// metadata.json file and whose name is a 16-hex part id (an on-disk part dir).
 func findPartDirs(root string) []string {
 	var dirs []string
 	_ = filepath.WalkDir(root, func(path string, d fs.DirEntry, err error) error {
